moved the remaining KVM functionality from linux/uhyve to linux/x86_64/kvm_cpu
