+++ src/linux/uhyve.rs
@@ -2,7 +2,6 @@
 //! create a Virtual Machine and load the kernel.
 
 use std::{
-	cmp,
 	ffi::OsString,
 	fmt,
 	path::{Path, PathBuf},
@@ -11,25 +10,19 @@ use std::{
 };
 
 use hermit_entry::boot_info::RawBootInfo;
-use kvm_bindings::*;
-use kvm_ioctls::VmFd;
-use log::debug;
-use vmm_sys_util::eventfd::EventFd;
 
 use crate::{
-	consts::*,
-	linux::{mem::MmapMemory, virtio::*, x86_64::kvm_cpu::KvmCpu, KVM},
+	linux::{
+		mem::MmapMemory,
+		virtio::*,
+		x86_64::kvm_cpu::{initialize_kvm, KvmCpu},
+	},
 	params::Params,
 	vm::{Vm, VmGuestMemory},
 	HypervisorResult,
 };
 
-const KVM_32BIT_MAX_MEM_SIZE: usize = 1 << 32;
-const KVM_32BIT_GAP_SIZE: usize = 768 << 20;
-const KVM_32BIT_GAP_START: usize = KVM_32BIT_MAX_MEM_SIZE - KVM_32BIT_GAP_SIZE;
-
 pub struct Uhyve {
-	vm: VmFd,
 	offset: u64,
 	entry_point: u64,
 	stack_address: u64,
@@ -62,11 +55,7 @@ impl Uhyve {
 	pub fn new(kernel_path: PathBuf, params: Params) -> HypervisorResult<Uhyve> {
 		let memory_size = params.memory_size.get();
 
-		let vm = KVM.create_vm()?;
-
 		let mem = MmapMemory::new(0, memory_size, 0, params.thp, params.ksm);
-
-		let sz = cmp::min(memory_size, KVM_32BIT_GAP_START);
 
 		// create virtio interface
 		// TODO: Remove allow once fixed:
@@ -74,80 +63,7 @@ impl Uhyve {
 		#[allow(clippy::arc_with_non_send_sync)]
 		let virtio_device = Arc::new(Mutex::new(VirtioNetPciDevice::new()));
 
-		let kvm_mem = kvm_userspace_memory_region {
-			slot: 0,
-			flags: mem.flags,
-			memory_size: sz as u64,
-			guest_phys_addr: mem.guest_address as u64,
-			userspace_addr: mem.host_address as u64,
-		};
-
-		unsafe { vm.set_user_memory_region(kvm_mem) }?;
-
-		if memory_size > KVM_32BIT_GAP_START + KVM_32BIT_GAP_SIZE {
-			let kvm_mem = kvm_userspace_memory_region {
-				slot: 1,
-				flags: mem.flags,
-				memory_size: (memory_size - KVM_32BIT_GAP_START - KVM_32BIT_GAP_SIZE) as u64,
-				guest_phys_addr: (mem.guest_address + KVM_32BIT_GAP_START + KVM_32BIT_GAP_SIZE)
-					as u64,
-				userspace_addr: (mem.host_address + KVM_32BIT_GAP_START + KVM_32BIT_GAP_SIZE)
-					as u64,
-			};
-
-			unsafe { vm.set_user_memory_region(kvm_mem) }?;
-		}
-
-		debug!("Initialize interrupt controller");
-
-		// create basic interrupt controller
-		vm.create_irq_chip()?;
-
-		if params.pit {
-			vm.create_pit2(kvm_pit_config::default()).unwrap();
-		}
-
-		// enable x2APIC support
-		let mut cap: kvm_enable_cap = kvm_bindings::kvm_enable_cap {
-			cap: KVM_CAP_X2APIC_API,
-			flags: 0,
-			..Default::default()
-		};
-		cap.args[0] =
-			(KVM_X2APIC_API_USE_32BIT_IDS | KVM_X2APIC_API_DISABLE_BROADCAST_QUIRK).into();
-		vm.enable_cap(&cap)
-			.expect("Unable to enable x2apic support");
-
-		// currently, we support only system, which provides the
-		// cpu feature TSC_DEADLINE
-		let mut cap: kvm_enable_cap = kvm_bindings::kvm_enable_cap {
-			cap: KVM_CAP_TSC_DEADLINE_TIMER,
-			..Default::default()
-		};
-		cap.args[0] = 0;
-		vm.enable_cap(&cap)
-			.expect_err("Processor feature `tsc deadline` isn't supported!");
-
-		let cap: kvm_enable_cap = kvm_bindings::kvm_enable_cap {
-			cap: KVM_CAP_IRQFD,
-			..Default::default()
-		};
-		vm.enable_cap(&cap)
-			.expect_err("The support of KVM_CAP_IRQFD is currently required");
-
-		let mut cap: kvm_enable_cap = kvm_bindings::kvm_enable_cap {
-			cap: KVM_CAP_X86_DISABLE_EXITS,
-			flags: 0,
-			..Default::default()
-		};
-		cap.args[0] =
-			(KVM_X86_DISABLE_EXITS_PAUSE | KVM_X86_DISABLE_EXITS_MWAIT | KVM_X86_DISABLE_EXITS_HLT)
-				.into();
-		vm.enable_cap(&cap)
-			.expect("Unable to disable exists due pause instructions");
-
-		let evtfd = EventFd::new(0).unwrap();
-		vm.register_irqfd(&evtfd, UHYVE_IRQ_NET)?;
+		initialize_kvm(&mem, params.pit)?;
 
 		let cpu_count = params.cpu_count.get();
 
@@ -157,7 +73,6 @@ impl Uhyve {
 		);
 
 		let mut hyve = Uhyve {
-			vm,
 			offset: 0,
 			entry_point: 0,
 			stack_address: 0,
@@ -219,14 +134,13 @@ impl Vm for Uhyve {
 	}
 
 	fn create_cpu(&self, id: u32) -> HypervisorResult<KvmCpu> {
-		Ok(KvmCpu::new(
+		KvmCpu::new(
 			id,
 			self.path.clone(),
 			self.args.clone(),
-			self.vm.create_vcpu(id.into())?,
 			self.mem.host_address,
 			self.virtio_device.clone(),
-		))
+		)
 	}
 
 	fn set_boot_info(&mut self, header: *const RawBootInfo) {
